added confirmation window to drop operation

+++ src/gui/DatabaseContextMenu.java
@@ -6,6 +6,7 @@ import javafx.event.EventHandler;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ContextMenu;
 import javafx.scene.control.MenuItem;
 import javafx.stage.Stage;
@@ -41,14 +42,20 @@ public class DatabaseContextMenu extends ContextMenu {
         deleteTableMenu.setOnAction(new EventHandler<ActionEvent>() {
             @Override
             public void handle(ActionEvent actionEvent) {
-                Client.getClient().write("DROP DATABASE " + databaseName + "\n");
-                String answer = Client.getClient().readLine();
+                Alert dialog = new Alert(Alert.AlertType.CONFIRMATION, "Drop database [" + databaseName + "]?",
+                        ButtonType.YES, ButtonType.CANCEL);
+                dialog.showAndWait();
 
-                if (!answer.equals("OK")) {
-                    Alert alert = new Alert(Alert.AlertType.ERROR, answer);
-                    alert.setTitle("Error");
-                    alert.setHeaderText(null);
-                    alert.showAndWait();
+                if (dialog.getResult() == ButtonType.YES) {
+                    Client.getClient().write("DROP DATABASE " + databaseName + "\n");
+                    String answer = Client.getClient().readLine();
+
+                    if (!answer.equals("OK")) {
+                        Alert alert = new Alert(Alert.AlertType.ERROR, answer);
+                        alert.setTitle("Error");
+                        alert.setHeaderText(null);
+                        alert.showAndWait();
+                    }
                 }
             }
         });
